src/AzureMarketplace/requirements.txt
azure-cosmosdb-table
msal
azure-core
azure-identity
msrest
azuremarketplace